requirements.txt
requests
pydantic
crawl4ai
google-generativeai
python-dotenv
streamlit
beautifulsoup4
playwright